+++ docs/conf.py
@@ -31,6 +31,7 @@ import os
 import sys
 import datetime
 from importlib import import_module
+import tomllib
 
 try:
     from sphinx_astropy.conf.v1 import *  # noqa
@@ -39,12 +40,6 @@ except ImportError:
     sys.exit(1)
 
 # Get configuration information from pyproject.toml
-
-try:
-    import tomllib
-except ImportError:
-    # FIXME: remove when we require Python >= 3.11
-    import tomli as tomllib
 
 with open(os.path.join(
         os.path.dirname(__file__), '..', 'pyproject.toml'), 'rb') as f: